to various files added static_cast<double>() to demonstrate explicit coersion

+++ Chapter2_1_17_18/Program2_1.h
@@ -5,6 +5,6 @@
 // <return type> nameOfFunction (<formal parameter list>);
 int computeSum();
 void addTwoInts();
-int average(int, int, int);
+double average(int, int, int);
 
 #endif
